// frontend/src/types/apiResponseVoid.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ApiResponseVoidErrors } from './apiResponseVoidErrors';

export interface ApiResponseVoid {
  code?: number;
  message?: string;
  data?: unknown;
  errors?: ApiResponseVoidErrors;
  timestamp?: string;
}